+++ src/main/java/com/bitbank/dto/MfaDTO.java
@@ -1,5 +1,7 @@
 package com.bitbank.dto;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
+
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
@@ -7,14 +9,15 @@ import lombok.Data;
 @Data
 public class MfaDTO {
 
-	public MfaDTO(String id, String code) {
-		this.id = id;
+	public MfaDTO(String tempSecret, String code) {
+		this.tempSecret = tempSecret;
 		this.code = code;
 	}
 
+	@JsonProperty("castle_id")
 	@NotNull(message = "has to be present")
 	@NotEmpty(message = "has to be not empty")
-	private String id;
+	private String tempSecret;
 
 	@NotNull(message = "has to be present")
 	@NotEmpty(message = "has to be not empty")