+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-# 1.1.0
+# 1.2.0
++ Change OTP library
++ Add EntityNotFoundException
++ Add method to store a temp secret
++ Generate a secret store
++ Login via temp secret
+
+## 1.1.0
 + Upgrade dependencies
 + Fix deprecated CORS
 + Add column `secret_mfa` to database